+++ common/cpp/AudioBuffer/android/AudioBufferWrapper.cpp
@@ -31,11 +31,11 @@ int AudioBufferWrapper::getNumberOfChannels() const {
   return numberOfChannels;
 }
 
-int16_t **AudioBufferWrapper::getChannelData(int channel) const {
+float **AudioBufferWrapper::getChannelData(int channel) const {
   return audioBuffer_->getChannelData(channel);
 }
 
-void AudioBufferWrapper::setChannelData(int channel, int16_t **data) const {
+void AudioBufferWrapper::setChannelData(int channel, float **data) const {
   audioBuffer_->setChannelData(channel, data);
 }
 } // namespace audioapi
